Refactor send email functionality
Sending email now is a separate service
Password reset logic moved to password controller
